fermat_mod_square/shift_word: widen SW_DIV_N's sw*nwt product before dividing
SW_DIV_N (fermat_mod_square.c) and its shift_word() sibling sw_div_n
(Mlucas.c) are both computed as sw*nwt/n with sw and nwt declared
uint32. The ASSERT immediately above the fermat_mod_square.c site
already casts to 64 bits (ASSERT((uint64)sw*nwt % n == 0, ...)) but
the assignment on the next line does not, so the product itself can
silently wrap mod 2^32 before the division ever sees it.

SW_DIV_N is the per-digit stride (mod nwt) into the IBDWT weight
table for the Fermat-mod forward/inverse weighting and the carry
routines' index setup, so a wrapped value applies the wrong 2^(k/nwt)
weight to nearly every digit.

Enumerated every (Fermat number, FFT length) pair where sw*nwt >= 2^32
(all Fermat-supported odd parts 7/15/63, table lengths up to 512M,
13-18 bits/digit):

  F31 @ 129024K (63*2^21, 16.25 bits/digit): correct SW_DIV_N 47, wrapped 14
  F32 @ 258048K (63*2^22, 16.25 bits/digit): correct SW_DIV_N 47, wrapped 14
  F33 @ 491520K (15*2^25, 17.07 bits/digit): correct SW_DIV_N 14, wrapped 5
  F33 @ 516096K (63*2^23, 16.25 bits/digit): correct SW_DIV_N 47, wrapped 6

All four lengths are already in get_fft_radices.c's table. None is a
default FFT length for its Fermat number (those are the neighbouring
powers of 2), so today they're reached only via explicit -fft, and at
those particular (F-number, length) pairs the 63-family/15-family
under-sizing already halts the run with a roundoff error or carry
overflow before the wrap could be observed end-to-end -- confirmed
identical behavior (ERR_CARRY at iter 6) between the unfixed and fixed
trees at F31@129024K. So this is latent today, not a live wrong
answer, but it stops being latent the moment either those lengths
become usable (HIACC carry, relaxed ROE limit) or the code is reused
elsewhere.

Fix: (uint32)((uint64)sw*nwt/n) at both sites, matching the cast the
neighbouring ASSERT already uses.

Verified unaffected lengths are bit-for-bit unchanged: -f 19 -fft 32
and -f 21 -fft 240, 100 iterations, avx2, give the same reference
Res64 (529E54642A813995 / 1DE0171591038250) before and after.

(cherry picked from commit 4dfaf1e0917e38f5882c1dad5dcd15b87d1a1808)

+++ src/Mlucas.c
@@ -4040,7 +4040,7 @@ uint64 	shift_word(double a[], int n, const uint64 p, const uint64 shift, const 
 			by initing a single DWT weight = 1.0 in the power-of-2 case and = 2^((j%nwt)/n) otherwise:
 			*/
 			nwt = (n >> trailz32(n));
-			sw_div_n = sw*nwt/n;
+			sw_div_n = (uint32)((uint64)sw*nwt/n);	// sw,nwt are uint32; sw*nwt can exceed 2^32 at the largest table lengths, so widen before dividing
 		}
 		else
 			ASSERT(TRANSFORM_TYPE == REAL_WRAPPER,"Require TRANSFORM_TYPE == REAL_WRAPPER");
